requirements.txt
aioredis==1.3.1
asgiref==3.2.7
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
beautifulsoup4==4.9.0
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
dj-static==0.0.6
Django==3.0.6
django-bootstrap4==1.1.1
django-channels==0.7.0
django-debug-toolbar==2.2
django-stdimage==5.1.1
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
msgpack==0.6.2
oauthlib==3.1.0
Pillow==7.1.2
progressbar2==3.51.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-utils==2.4.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
static3==0.7.0
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
zope.interface==5.1.0
